+++ projects/summer-fi/constants.js
@@ -1,0 +1,7 @@
+const creationBlocks = {
+  AutomationV1Contract: 14583413,
+};
+
+module.exports = {
+  creationBlocks,
+};
