+++ projects/summer-fi/contracts.js
@@ -1,0 +1,6 @@
+module.exports = {
+  AutomationV1Contract: "0x6E87a7A0A03E51A741075fDf4D1FCce39a4Df01b",
+  McdView: "0x6FBef01dd5B8e85b19733AB2A8e243FA3870623B",
+  CdpManager: "0x5ef30b9986345249bc32d8928b7ee64de9435e39",
+  IlkRegistry: "0x5a464c28d19848f44199d003bef5ecc87d090f87",
+};
