+++ projects/summer-fi/dpm-positions.js
@@ -1,0 +1,7 @@
+const dpmPositions = async ({ api }) => {
+  // do something
+};
+
+module.exports = {
+  dpmPositions,
+};
