+++ projects/summer-fi/logs-topic.js
@@ -1,0 +1,8 @@
+module.exports = {
+  TriggerAdded:
+    "0xcb616360dd177f28577e33576c8ac7ffcc1008cba7ac2323e0b2f170faf60bd2",
+  TriggerExecuted:
+    "0xc10f224f2f1ceab5e36f97effaa05c4b75eccbecd77b65bfb20c484de9096cdd",
+  TriggerRemoved:
+    "0xb4a1fc324bd863f8cd42582bebf2ce7f2d309c6a84bf371f28e069f95a4fa9e1",
+};
